RW05-1000-10p with ISPN53, FFLB and FFLF
Note: In FFLB the timers are not starting at the same time because of the
unfair locking mechanism.  So, I should add a synchronization barrier (between
populating the keys and starting the timer) to ensure the correct behavior.

Also, FFLB, appears to be constantly under ISPN53.  Next, I'll try
RW01-100-10p in which ISPN52 can improve whereas FFLB-ISPN52 cannot.  I expect
FFLF-ISPN53 to be able to improve as long ISPN53 does.  If not, I should, at
the very least, run FFLF-ISPN52 (*AND* ISPN52 just to make sure that it
actually improves).