FutureGrid results. The latest are quite good.
Results in
auto-results.20131121-114900-FF@8246873-RG@56a4b01-indexingReadset-writeset1-withCommitStats
use:

  - print commit stats
  - indexed read set for intersections
  - write set size = 1
  - radargun bug with 'inTransaction' fixed.

Supposedly the last item accounts for the reduction in performance in
RW10-100-1p and RW20-100-1p when compared to previous runs.  I re-run the
tests without the first two items and the results were the same.  The only
delta was thus on the write set size and the 'inTransaction' fix.  While the
write set size could affect the results (higher rate of transactions?), and
fairly confident that the bug fix is the explanation.